refactor(web): move SystemStatus and LanguageSelector out of Layout.tsx
Layout.tsx held three unrelated components; the sidebar health pill and
the language picker now live in their own files. Pure move, no behaviour
change; the existing Layout tests still cover both through the sidebar.
Layout.tsx drops from 1583 to 1065 lines and its allowlist entry ratchets
down to match.
