+++ web/src/components/Layout.tsx
@@ -11,7 +11,6 @@ import {
 	GitCompare,
 	History as HistoryIcon,
 	KeyRound,
-	Languages,
 	LayoutDashboard,
 	MessageSquare,
 	MessagesSquare,
@@ -42,534 +41,17 @@ import { useIdleLogout } from "../hooks/useIdleLogout";
 import { useManaged } from "../hooks/useManaged";
 import { useReadOnly } from "../hooks/useReadOnly";
 import { useRefreshDiscoveryBadge } from "../hooks/useRefreshDiscoveryBadge";
-import i18next, { LANGUAGE_STORAGE_KEY } from "../i18n";
 import { useDiscoveryRetest } from "../pages/Providers/useDiscoveryRetest";
-import { CollapsibleToggle, useCollapsible } from "./CollapsibleToggle";
 import { ConfirmDialog } from "./ConfirmDialog";
-import { CountryFlag } from "./CountryFlag";
 import { ErrorShelf } from "./ErrorShelf";
+import { LanguageSelector } from "./LanguageSelector";
 import { Logo } from "./Logo";
 import { ModelDiscrepancyModal } from "./ModelDiscrepancyModal";
 import { ProviderQuotaPanel } from "./ProviderQuotaPanel";
-import {
-	formatCount,
-	formatMemoryMB,
-	formatThroughput,
-	formatUptime,
-	unitClass,
-} from "./systemStatusFormat";
-
-function SystemStatus() {
-	const { t } = useTranslation();
-	const {
-		data: stats,
-		isError,
-		dataUpdatedAt,
-	} = useQuery({
-		queryKey: ["system"],
-		queryFn: () => api.system.get(),
-		refetchInterval: 10000,
-		staleTime: 3000,
-		retry: 1,
-	});
-
-	const { collapsed, toggle: toggleCollapsed } = useCollapsible(
-		"sidebarStatsCollapsed",
-	);
-
-	const app = stats?.app;
-	// HA fleet membership. Present only while Front Desk is in contact; a
-	// standalone instance has no `fleet` block and renders no HA line.
-	const fleet = stats?.fleet;
-	const haColor = fleet
-		? {
-				primary: "text-green-400",
-				member: "text-green-400",
-				warning: "text-orange-400",
-				member_sync_blocked: "text-red-400",
-			}[fleet.state]
-		: "";
-	const haValue = fleet
-		? {
-				primary: t("layout.ha.primary"),
-				member: t("layout.ha.member"),
-				warning: t("layout.ha.warning"),
-				member_sync_blocked: t("layout.ha.error"),
-			}[fleet.state]
-		: "";
-	const haTooltip = fleet
-		? {
-				primary: t("layout.ha.tooltips.primary"),
-				member: fleet.primary_name
-					? t("layout.ha.tooltips.memberFrom", { name: fleet.primary_name })
-					: t("layout.ha.tooltips.member"),
-				warning: t("layout.ha.tooltips.warning"),
-				member_sync_blocked: t("layout.ha.tooltips.error"),
-			}[fleet.state]
-		: "";
-	const docker = stats?.docker;
-	const inContainer = app?.in_container;
-	const hasLimit = !!(inContainer && app?.memory_limit_bytes);
-
-	const useDocker = docker?.available;
-
-	const cpuPct = useDocker ? docker.cpu_percent : app?.cpu_percent;
-	const procs = useDocker ? docker.procs : app?.procs;
-	const netRx = useDocker ? docker.net_rx_bytes_sec : app?.net_rx_bytes_sec;
-	const netTx = useDocker ? docker.net_tx_bytes_sec : app?.net_tx_bytes_sec;
-	const diskRead = useDocker
-		? docker.disk_read_bytes_sec
-		: app?.disk_read_bytes_sec;
-	const diskWrite = useDocker
-		? docker.disk_write_bytes_sec
-		: app?.disk_write_bytes_sec;
-
-	const dc = (v: number | undefined, w: number, c: number, inv?: boolean) => {
-		if (v == null) return "";
-		const bad = inv ? v <= c : v >= c;
-		const warn = inv ? v <= w : v >= w;
-		return bad ? "text-red-400" : warn ? "text-orange-400" : "";
-	};
-
-	// A cache-hit sample over a near-idle window (one autovacuum scan, a stray
-	// catalog read) says nothing about cache health, so below this many block
-	// accesses the hit cell shows a dash instead of colour-coding noise. The
-	// backend omits cache_window_blocks entirely for no-window samples.
-	const cacheHitLive = (stats?.db?.cache_window_blocks ?? 0) >= 1000;
-
-	const dockerMem = useDocker && docker.memory_limit_bytes > 0;
-	const memUsagePct = dockerMem
-		? (docker.memory_usage_bytes / docker.memory_limit_bytes) * 100
-		: hasLimit && app?.memory_limit_bytes
-			? (app.memory_current_bytes / app.memory_limit_bytes) * 100
-			: undefined;
-	const appMem = dockerMem ? (
-		<>
-			{formatMemoryMB(docker.memory_usage_bytes / 1024 / 1024)} /{" "}
-			{formatMemoryMB(docker.memory_limit_bytes / 1024 / 1024)}
-		</>
-	) : hasLimit ? (
-		<>
-			{formatMemoryMB(app.memory_current_bytes / 1024 / 1024)} /{" "}
-			{formatMemoryMB(app.memory_limit_bytes / 1024 / 1024)}
-		</>
-	) : app ? (
-		<>
-			{formatMemoryMB(app.heap_alloc_mb)}
-			<span className={unitClass}> {t("layout.stats.heap")}</span>
-		</>
-	) : (
-		"-"
-	);
-
-	const dash = <span className="text-(--text-muted)">-</span>;
-
-	return (
-		<div className="sidebar-stats-pill">
-			<div className="sidebar-stats-trigger">
-				<div
-					className="flex justify-between items-center text-[11px] font-mono text-(--text-tertiary) flex-1 min-w-0"
-					title={t("layout.status.apiStatus")}
-				>
-					<span>{t("layout.status.apiStatus")}</span>
-					<span
-						className={`flex items-center ${isError ? "text-red-400" : "text-green-400"}`}
-					>
-						<span
-							// Remount on each successful refetch so the online dot
-							// replays its one-shot pulse; offline keeps a constant key
-							// so its looping pulse runs uninterrupted.
-							key={isError ? "offline" : dataUpdatedAt}
-							className={`w-1.5 h-1.5 rounded-full mr-1.5 ${isError ? "bg-red-400 status-dot-offline" : "bg-green-400 status-dot-online"}`}
-						/>
-						{isError ? t("layout.status.error") : t("layout.status.online")}
-					</span>
-				</div>
-			</div>
-			<div
-				className={`grid transition-[grid-template-rows] duration-300 ease-in-out ${collapsed ? "grid-rows-[0fr]" : "grid-rows-[1fr]"}`}
-			>
-				<div className="overflow-hidden">
-					<div className="sidebar-stats-content space-y-0.5 text-[11px] font-mono system-status">
-						{/* HA fleet membership (only while managed by a Front Desk) */}
-						{fleet && (
-							<div
-								className="flex justify-between items-center text-(--text-tertiary)"
-								title={haTooltip}
-								data-testid="ha-status"
-							>
-								<span>HA</span>
-								<span className={haColor}>{haValue}</span>
-							</div>
-						)}
-
-						{/* Uptime */}
-						<div
-							className="flex justify-between items-center text-(--text-tertiary)"
-							title={t("layout.tooltips.uptime")}
-						>
-							<span>{t("layout.stats.uptime")}</span>
-							<span className="text-(--text-secondary)">
-								{app ? formatUptime(app.uptime_seconds) : dash}
-							</span>
-						</div>
-
-						{/* CPU + Processes */}
-						<div
-							className="flex justify-between items-center text-(--text-tertiary)"
-							title={
-								useDocker
-									? t("layout.stats.aggregateCpu", {
-											count: docker.container_count,
-										})
-									: t("layout.stats.cpu")
-							}
-						>
-							<span>{t("layout.stats.cpu")}</span>
-							<span className={`text-(--text-secondary) ${dc(cpuPct, 75, 90)}`}>
-								{cpuPct != null && cpuPct >= 0 ? (
-									<>
-										<span>
-											{cpuPct.toFixed(1)}
-											<span className={unitClass}>%</span>
-										</span>
-										{procs != null && procs > 0 && (
-											<>
-												<span className="text-(--text-secondary) mx-1">|</span>
-												<span>
-													{procs}
-													<span className={unitClass}>
-														{" "}
-														{t("layout.stats.procs", { count: procs })}
-													</span>
-												</span>
-											</>
-										)}
-									</>
-								) : (
-									dash
-								)}
-							</span>
-						</div>
-
-						{/* Network */}
-						<div
-							className="flex justify-between items-center text-(--text-tertiary)"
-							title={
-								useDocker
-									? t("layout.stats.aggregateNetwork", {
-											count: docker.container_count,
-										})
-									: t("layout.stats.network")
-							}
-						>
-							<span>{t("layout.stats.network")}</span>
-							<span className="text-(--text-secondary) tabular-nums">
-								<span className="text-sky-400/60 inline-block min-w-22 text-right">
-									{typeof netRx === "number" ? (
-										<>↓{formatThroughput(netRx)}</>
-									) : (
-										dash
-									)}
-								</span>
-								<span className="text-amber-400/60 inline-block min-w-22 text-right">
-									{typeof netTx === "number" ? (
-										<>↑{formatThroughput(netTx)}</>
-									) : (
-										dash
-									)}
-								</span>
-							</span>
-						</div>
-
-						{/* Disk I/O */}
-						<div
-							className="flex justify-between items-center text-(--text-tertiary)"
-							title={
-								useDocker
-									? t("layout.stats.aggregateDisk", {
-											count: docker.container_count,
-										})
-									: t("layout.stats.disk")
-							}
-						>
-							<span>{t("layout.stats.disk")}</span>
-							<span className="text-(--text-secondary) tabular-nums">
-								<span className="text-sky-400/60 inline-block min-w-22 text-right">
-									{typeof diskRead === "number" ? (
-										<>↓{formatThroughput(diskRead)}</>
-									) : (
-										dash
-									)}
-								</span>
-								<span className="text-amber-400/60 inline-block min-w-22 text-right">
-									{typeof diskWrite === "number" ? (
-										<>↑{formatThroughput(diskWrite)}</>
-									) : (
-										dash
-									)}
-								</span>
-							</span>
-						</div>
-
-						{/* Memory */}
-						<div
-							className="flex justify-between items-center text-(--text-tertiary)"
-							title={
-								dockerMem
-									? t("layout.stats.aggregateMemory", {
-											count: docker.container_count,
-										})
-									: t("layout.stats.memory")
-							}
-						>
-							<span>{t("layout.stats.memory")}</span>
-							<span
-								className={`text-(--text-secondary) ${dc(memUsagePct, 75, 90)}`}
-							>
-								{app ? appMem : dash}
-							</span>
-						</div>
-
-						{/* Goroutines */}
-						<div
-							className="flex justify-between items-center text-(--text-tertiary)"
-							title={t("layout.tooltips.goroutines")}
-						>
-							<span>{t("layout.stats.goroutines")}</span>
-							<span
-								className={`text-(--text-secondary) ${dc(app?.goroutines, 300, 1000)}`}
-							>
-								{app ? app.goroutines.toLocaleString() : dash}
-							</span>
-						</div>
-
-						{/* Requests Today */}
-						<div
-							className="flex justify-between items-center text-(--text-tertiary)"
-							title={t("layout.tooltips.requestsToday")}
-						>
-							<span>{t("layout.stats.requestsToday")}</span>
-							<span className="text-(--text-secondary)">
-								{app && app.requests_today > 0
-									? formatCount(app.requests_today)
-									: dash}
-							</span>
-						</div>
-
-						{/* DB: size & hit ratio / connections & tx/sec */}
-						<div className="flex justify-between items-center text-(--text-tertiary)">
-							<span className="self-center">{t("layout.stats.db")}</span>
-							<span className="grid grid-cols-[1fr_auto_1fr] grid-rows-[auto_auto] gap-x-2 items-center text-right">
-								{stats?.db ? (
-									<>
-										<span
-											className="text-(--text-secondary)"
-											title={t("layout.tooltips.dbSize")}
-										>
-											{formatMemoryMB(stats.db.size_mb)}
-										</span>
-										<span className="text-(--text-secondary)">|</span>
-										<span
-											className={`text-(--text-secondary) ${cacheHitLive ? dc(stats.db.cache_hit_ratio, 90, 80, true) : ""}`}
-											title={t("layout.tooltips.dbHitRatio")}
-										>
-											{t("layout.stats.hit")}{" "}
-											{cacheHitLive ? (
-												<>
-													{stats.db.cache_hit_ratio}
-													<span className={unitClass}>%</span>
-												</>
-											) : (
-												dash
-											)}
-										</span>
-										<span
-											className="text-(--text-secondary)"
-											title={t("layout.tooltips.dbConnections")}
-										>
-											{stats.db.connections}
-											<span className={unitClass}>
-												{" "}
-												{t("layout.stats.conn")}
-											</span>
-										</span>
-										<span className="text-(--text-secondary)">|</span>
-										<span
-											className="text-(--text-secondary)"
-											title={t("layout.tooltips.dbTxPerSec")}
-										>
-											{stats.db.tx_per_sec.toFixed(1)}
-											<span className={unitClass}>
-												{" "}
-												{t("layout.stats.txPerSec")}
-											</span>
-										</span>
-									</>
-								) : (
-									<>
-										<span className="text-(--text-muted)">-</span>
-										<span className="text-(--text-secondary)">|</span>
-										<span className="text-(--text-muted)">-</span>
-										<span className="text-(--text-muted)">-</span>
-										<span className="text-(--text-secondary)">|</span>
-										<span className="text-(--text-muted)">-</span>
-									</>
-								)}
-							</span>
-						</div>
-					</div>
-				</div>
-			</div>
-			<div className="sidebar-stats-footer">
-				<CollapsibleToggle
-					collapsed={collapsed}
-					onToggle={toggleCollapsed}
-					size={10}
-					iconStyle="double"
-					expandTitle={t("layout.expandStats")}
-					collapseTitle={t("layout.collapseStats")}
-				/>
-			</div>
-		</div>
-	);
-}
+import { SystemStatus } from "./SystemStatus";
 
 interface LayoutProps {
 	children: React.ReactNode;
-}
-
-// Language names are autonyms (each language in its own script), shown
-// identically in every UI locale — the industry standard for language pickers,
-// so a user stranded in the wrong language can still recognize their own.
-// English is intentionally last so it sits at the bottom of the upward-opening
-// menu (nearest the trigger) in every locale.
-const SUPPORTED_LANGUAGES = [
-	{ code: "af", label: "Afrikaans" },
-	{ code: "ar", label: "العربية" },
-	{ code: "ca", label: "Català" },
-	{ code: "cs", label: "Čeština" },
-	{ code: "da", label: "Dansk" },
-	{ code: "de", label: "Deutsch" },
-	{ code: "el", label: "Ελληνικά" },
-	{ code: "es", label: "Español" },
-	{ code: "fi", label: "Suomi" },
-	{ code: "fr", label: "Français" },
-	{ code: "he", label: "עברית" },
-	{ code: "hu", label: "Magyar" },
-	{ code: "it", label: "Italiano" },
-	{ code: "ja", label: "日本語" },
-	{ code: "ko", label: "한국어" },
-	{ code: "nl", label: "Nederlands" },
-	{ code: "no", label: "Norsk" },
-	{ code: "pl", label: "Polski" },
-	{ code: "pt", label: "Português" },
-	{ code: "ro", label: "Română" },
-	{ code: "ru", label: "Русский" },
-	{ code: "sk", label: "Slovenčina" },
-	{ code: "sr", label: "Српски" },
-	{ code: "sv", label: "Svenska" },
-	{ code: "tr", label: "Türkçe" },
-	{ code: "uk", label: "Українська" },
-	{ code: "vi", label: "Tiếng Việt" },
-	{ code: "zh", label: "中文" },
-	{ code: "en", label: "English" },
-] as const;
-
-function LanguageSelector() {
-	const { t, i18n } = useTranslation();
-	const [open, setOpen] = useState(false);
-	const ref = useRef<HTMLDivElement>(null);
-	const scrollRef = useRef<HTMLDivElement>(null);
-
-	// Set document direction for RTL languages
-	useEffect(() => {
-		const rtlLanguages = new Set(["ar", "he"]);
-		const lang = i18n.resolvedLanguage as string;
-		document.documentElement.dir = rtlLanguages.has(lang) ? "rtl" : "ltr";
-	}, [i18n.resolvedLanguage]);
-
-	useEffect(() => {
-		function handleClickOutside(e: MouseEvent) {
-			if (ref.current && !ref.current.contains(e.target as Node)) {
-				setOpen(false);
-			}
-		}
-		if (open) {
-			document.addEventListener("mousedown", handleClickOutside);
-			return () =>
-				document.removeEventListener("mousedown", handleClickOutside);
-		}
-	}, [open]);
-
-	// Scroll the active language into view when dropdown opens
-	useEffect(() => {
-		if (open && scrollRef.current) {
-			const active = scrollRef.current.querySelector("[aria-selected='true']");
-			active?.scrollIntoView({ block: "nearest" });
-		}
-	}, [open]);
-
-	if (SUPPORTED_LANGUAGES.length <= 1) return null;
-
-	return (
-		<div ref={ref} className="relative">
-			<button
-				type="button"
-				onClick={() => setOpen((v) => !v)}
-				className="sidebar-footer-link text-gray-400 hover:text-white ui-btn hover:bg-white/5"
-				title={t("layout.language.label")}
-				aria-label={t("layout.language.label")}
-				data-testid="language-trigger"
-			>
-				<Languages size={14} strokeWidth={2} />
-			</button>
-			{open && (
-				// Outer wrapper owns the rounding + border and clips its overflow so
-				// the inner scrollbar stays inside the rounded corners instead of
-				// painting over them. The scroll lives on the inner element.
-				<div className="ui-popover absolute bottom-full left-1/2 -translate-x-1/2 mb-1 min-w-[120px] bg-gray-800 border border-gray-700 rounded-(--radius-card) shadow-lg z-50 overflow-hidden">
-					<div
-						ref={scrollRef}
-						className="py-1 max-h-[50vh] overflow-y-auto overscroll-contain"
-						role="listbox"
-					>
-						{SUPPORTED_LANGUAGES.map((lang) => (
-							<button
-								key={lang.code}
-								type="button"
-								role="option"
-								aria-selected={
-									(i18n.resolvedLanguage ?? i18n.language) === lang.code
-								}
-								data-testid={`language-option-${lang.code}`}
-								id={`language-option-${lang.code}`}
-								onClick={() => {
-									i18next.changeLanguage(lang.code);
-									// Persist every deliberate choice — including English —
-									// so the effective priority is strictly
-									// user choice > system locale > English. The browser
-									// locale is never auto-cached (caches: [] in
-									// i18n/index.ts), so an explicit pick always wins on
-									// the next visit until the user changes it again.
-									localStorage.setItem(LANGUAGE_STORAGE_KEY, lang.code);
-									setOpen(false);
-								}}
-								className={`w-full text-left px-3 py-1.5 text-xs transition-colors flex items-center gap-1.5 ${
-									(i18n.resolvedLanguage ?? i18n.language) === lang.code
-										? "text-white bg-white/10"
-										: "text-gray-400 hover:text-white hover:bg-white/5"
-								}`}
-							>
-								<CountryFlag code={lang.code} />
-								{lang.label}
-							</button>
-						))}
-					</div>
-				</div>
-			)}
-		</div>
-	);
 }
 
 // ReadOnlyBanner is shown on every page when the server runs in read-only
